backend/requirements.txt
APIFlask==2.0.2
Flask_Cors==3.0.10
numpy==1.22.4
pandas==2.1.3
pymssql==2.2.8
python_dateutil==2.8.2
tinydb==4.8.0
